fix(security): remove client secret enforcement and use proxy if configured
