+++ apps/desktop/src/routes/login/+page.svelte
@@ -9,9 +9,10 @@
   import { handleError } from "$lib/logger";
   import {
     getOAuthAvailability,
-    handleOAuthCallback,
     openOAuthAuthUrl,
     startOAuthFlow,
+    processOAuthCallback,
+    sanitizeErrorForAudit,
     type OAuthAvailability,
     type OAuthProvider,
   } from "$lib/oauth";
@@ -27,30 +28,6 @@
     const [, domain] = email.split("@");
     if (!domain) return "***@***";
     return `***@${domain}`;
-  }
-
-  // Helper function to sanitize error messages for audit logging
-  // Removes potentially sensitive data like URLs, tokens, or OAuth parameters
-  function sanitizeErrorForAudit(err: unknown): string {
-    const errorStr = String(err);
-    // Map specific errors to generic categories without leaking details
-    if (errorStr.includes("callback") || errorStr.includes("aroeira://")) {
-      return "callback_processing_error";
-    }
-    if (errorStr.includes("token") || errorStr.includes("exchange")) {
-      return "token_exchange_error";
-    }
-    if (errorStr.includes("network") || errorStr.includes("fetch")) {
-      return "network_error";
-    }
-    if (errorStr.includes("expired") || errorStr.includes("invalid")) {
-      return "session_invalid_or_expired";
-    }
-    if (errorStr.includes("denied") || errorStr.includes("cancelled")) {
-      return "user_denied_or_cancelled";
-    }
-    // Generic fallback - never log raw error content
-    return "authentication_error";
   }
 
   /**
@@ -95,182 +72,24 @@
   let policy = $state<PasswordPolicy>({ level: "secure", min_length: 8 });
   let oauthAvailability = $state<OAuthAvailability | null>(null);
 
-  // Serialize OAuth callback handling to avoid races without dropping events
-  let oauthCallbackQueue: Promise<void> = Promise.resolve();
-
   /**
-   * Process an OAuth callback URL from deep linking.
-   * This handles the aroeira://auth/callback URLs.
+   * Wrapper to process OAuth callback using the utility function with component state callbacks.
    */
-  function processOAuthCallback(rawUrl: string): Promise<void> {
-    // Only emit minimal info; never include code/state/raw URL
-    if (import.meta.env.DEV) {
-      console.debug("Processing OAuth callback", { length: rawUrl.length });
-    }
-    // Defensive bound to avoid processing extremely large deep-link payloads
-    if (rawUrl.length > 8192) {
-      resetOAuthState("Authentication callback was invalid. Please try again.");
-      return Promise.resolve();
-    }
-
-    let parsed: URL | null = null;
-    try {
-      parsed = new URL(rawUrl);
-    } catch {
-      // If this looks like our scheme but isn't parseable, treat as a failed callback
-      if (rawUrl.startsWith("aroeira:")) {
-        resetOAuthState(
-          "Authentication callback was invalid. Please try again.",
-        );
-      }
-      return Promise.resolve();
-    }
-
-    const isAroeiraProtocol =
-      parsed.protocol === "aroeira:" &&
-      parsed.hostname === "auth" &&
-      parsed.pathname === "/callback";
-
-    const isLocalhostDev =
-      import.meta.env.DEV &&
-      parsed.protocol === "http:" &&
-      parsed.hostname === "localhost" &&
-      parsed.pathname === "/auth/callback";
-
-    const isOAuthCallback = isAroeiraProtocol || isLocalhostDev;
-
-    if (import.meta.env.DEV) {
-      console.debug("Parsed URL", {
-        protocol: parsed.protocol,
-        hostname: parsed.hostname,
-        pathname: parsed.pathname,
-        isOAuthCallback,
-      });
-    }
-
-    if (!isOAuthCallback) {
-      // Ignore unrelated deep links; don't cancel an in-progress OAuth flow.
-      if (import.meta.env.DEV) {
-        console.debug("Not an OAuth callback, ignoring");
-      }
-      return Promise.resolve();
-    }
-
-    const callbackUrl = parsed;
-
-    oauthCallbackQueue = oauthCallbackQueue
-      .catch(() => {
-        // Keep the queue alive even if a previous callback failed
-      })
-      .then(async () => {
-        if (destroyed) return;
-
-        // Restore loading state from localStorage if not already set
-        if (!oauthLoading) {
-          const savedProvider = localStorage.getItem("oauth_pending_provider");
-          oauthLoading =
-            savedProvider === "google" || savedProvider === "github"
-              ? (savedProvider as OAuthProvider)
-              : null;
-
-          if (!oauthLoading && savedProvider) {
-            localStorage.removeItem("oauth_pending_provider");
-            localStorage.removeItem("oauth_pending_state");
-          }
-        }
-
-        // Check for OAuth provider errors (user denied/cancelled)
-        const oauthError = callbackUrl.searchParams.get("error");
-        if (oauthError) {
-          resetOAuthState(
-            "Authentication was cancelled or denied. Please try again.",
-          );
-          return;
-        }
-
-        // Validate callback contains authorization code
-        const code = callbackUrl.searchParams.get("code");
-        const state = callbackUrl.searchParams.get("state");
-        if (import.meta.env.DEV) {
-          console.debug("OAuth callback params", {
-            hasCode: !!code,
-            hasState: !!state,
-            codeLength: code?.length,
-            stateLength: state?.length,
-          });
-        }
-        if (!code) {
-          resetOAuthState(
-            "Authentication callback was invalid. Please try again.",
-          );
-          return;
-        }
-
-        // Expire stale OAuth pending state (> 2 minutes)
-        const startedAtStr = localStorage.getItem("oauth_pending_started_at");
-        const startedAt = startedAtStr ? Number(startedAtStr) : NaN;
-        const maxAgeMs = 2 * 60 * 1000;
-
-        if (!Number.isFinite(startedAt) || Date.now() - startedAt > maxAgeMs) {
-          resetOAuthState("Authentication session expired. Please try again.");
-          return;
-        }
-
-        // Validate state before invoking backend exchange
-        const pendingState = localStorage.getItem("oauth_pending_state");
-        const callbackState = callbackUrl.searchParams.get("state");
-        if (import.meta.env.DEV) {
-          console.debug("State validation", {
-            hasPendingState: !!pendingState,
-            hasCallbackState: !!callbackState,
-            match: pendingState === callbackState,
-          });
-        }
-        if (!pendingState || !callbackState || pendingState !== callbackState) {
-          resetOAuthState(
-            "Authentication session was invalid. Please try again.",
-          );
-          return;
-        }
-
-        error = "";
-
-        // Normalize localhost dev callback to aroeira:// scheme for backend consistency
-        const callbackForBackend = isLocalhostDev
-          ? `aroeira://auth/callback${parsed.search}`
-          : rawUrl;
-
-        try {
-          if (import.meta.env.DEV) {
-            console.debug("Calling handleOAuthCallback with backend", {
-              normalized: isLocalhostDev,
-            });
-          }
-          const user = await handleOAuthCallback(callbackForBackend);
-          console.log("OAuth callback successful, user:", {
-            provider: user.provider,
-            email: redactEmail(user.email),
-          });
-          logAuditEvent("oauth_login", true, {
-            provider: user.provider,
-            email: redactEmail(user.email),
-          });
-          setSessionId();
-          // Clean up OAuth state on success
-          resetOAuthState();
-          await goto(resolve("/dashboard"), { replaceState: true });
-        } catch (err: unknown) {
-          console.error("OAuth callback failed:", sanitizeErrorForAudit(err));
-          logAuditEvent("oauth_login", false, {
-            error: sanitizeErrorForAudit(err),
-          });
-          error = handleError(err, "OAuth authentication");
-        } finally {
-          resetOAuthState();
-        }
-      });
-
-    return oauthCallbackQueue;
+  function handleDeepLink(rawUrl: string): Promise<void> {
+    if (destroyed) return Promise.resolve();
+    return processOAuthCallback(
+      rawUrl,
+      {
+        setLoading: (p) => {
+          oauthLoading = p;
+        },
+        setError: (msg) => {
+          error = msg;
+        },
+        resetState: resetOAuthState,
+      },
+      () => oauthLoading,
+    );
   }
 
   onMount(async () => {
@@ -299,7 +118,7 @@
       const urls = await getCurrent();
       if (urls && urls.length > 0) {
         for (const url of urls) {
-          await processOAuthCallback(url);
+          await handleDeepLink(url);
         }
       }
     } catch (err) {
@@ -320,7 +139,7 @@
           });
         }
         for (const url of urls) {
-          await processOAuthCallback(url);
+          await handleDeepLink(url);
         }
       });
       if (import.meta.env.DEV) {
@@ -344,7 +163,7 @@
               length: event.payload?.length ?? 0,
             });
           }
-          await processOAuthCallback(event.payload);
+          await handleDeepLink(event.payload);
         },
       );
       if (import.meta.env.DEV) {
